+++ moderation/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-__version__ = '0.7.0'
+__version__ = '0.8.0'
 
 default_app_config = "moderation.apps.ModerationConfig"