+++ moderation/filterspecs.py
@@ -1,13 +1,13 @@
 from django.contrib.admin.filters import FieldListFilter
 from django.contrib.contenttypes.models import ContentType
 from django.utils.encoding import smart_text
-from django.utils.translation import ugettext as _
+from django.utils.translation import gettext as _
 
 from . import moderation
 
 
 def _registered_content_types():
-    "Return sorted content types for all registered models."
+    """Return sorted content types for all registered models."""
     content_types = []
     registered = list(moderation._registered_models.keys())
     registered.sort(key=lambda obj: obj.__name__)